+++ core/bone_utils.py
@@ -93,14 +93,17 @@ def align_armatures_by_name(source_arm, target_arm, mode='POS_ONLY', skip_fn=Non
     返回对齐骨骼数。执行后处于 OBJECT 模式，active 对象为 target_arm。
     """
     s_mat = source_arm.matrix_world
+    s_mat3 = s_mat.to_3x3()
     src_data = {}
     for b in source_arm.data.bones:
-        src_data[b.name] = (s_mat @ b.head_local.copy(), s_mat @ b.tail_local.copy())
+        x_world = (s_mat3 @ b.matrix.col[0].to_3d()).normalized()
+        src_data[b.name] = (s_mat @ b.head_local.copy(), s_mat @ b.tail_local.copy(), x_world)
 
     bpy.context.view_layer.update()
     bpy.context.view_layer.objects.active = target_arm
     bpy.ops.object.mode_set(mode='EDIT')
     t_mat_inv = target_arm.matrix_world.inverted()
+    t_mat3_inv = t_mat_inv.to_3x3()
 
     count = 0
     for b in target_arm.data.edit_bones:
@@ -108,12 +111,13 @@ def align_armatures_by_name(source_arm, target_arm, mode='POS_ONLY', skip_fn=Non
             continue
         if skip_fn and skip_fn(b.name):
             continue
-        src_head, src_tail = src_data[b.name]
+        src_head, src_tail, src_x_world = src_data[b.name]
         old_head = b.head.copy()
         new_head = t_mat_inv @ src_head
         if mode == 'FULL':
             b.head = new_head
             b.tail = t_mat_inv @ src_tail
+            b.align_roll(t_mat3_inv @ src_x_world)
         else:
             orig_vec = b.tail - b.head
             b.head = new_head
