+++ games/re9/__init__.py
@@ -1,13 +1,15 @@
-from . import operators
-from . import batch_export
-from . import batch_export_ui
+from . import operators, batch_export, batch_export_ui, mdf_tex_processor, mdf_tex_processor_ui
 
 def register():
     operators.register()
     batch_export.register()
     batch_export_ui.register()
+    mdf_tex_processor.register()
+    mdf_tex_processor_ui.register()
 
 def unregister():
+    mdf_tex_processor_ui.unregister()
+    mdf_tex_processor.unregister()
     batch_export_ui.unregister()
     batch_export.unregister()
     operators.unregister()
